channel/requirements.txt
requests==2.22.0
PyYAML==5.1.2
pytz==2019.2
pandas==0.25.1
google-cloud-storage==1.19.0
google-cloud-bigquery==1.20.0
